- Committing outstanding changes before I move on...

+++ flume-integration-tests/src/test/java/com/cloudera/flume/IntegrationTestSupport.java
@@ -1,7 +1,9 @@
 package com.cloudera.flume;
 
+import java.io.BufferedReader;
 import java.io.File;
 import java.io.IOException;
+import java.io.InputStreamReader;
 
 import org.apache.commons.io.FileUtils;
 import org.slf4j.Logger;
@@ -31,14 +33,47 @@ public class IntegrationTestSupport {
         "Flume distribution directory " + distributionDirectory
             + " doesn't exist or isn't a directory");
 
-    FileUtils.copyDirectory(distributionDirectory, workingDirectory);
+    try {
+      copy(distributionDirectory, workingDirectory);
+    } catch (InterruptedException e) {
+    }
   }
 
   public void tearDown() throws IOException {
     logger.info("Tearing down Flume integration test environment in {}",
         workingDirectory);
 
-    FileUtils.deleteDirectory(workingDirectory);
+    if (workingDirectory != null) {
+      FileUtils.deleteDirectory(workingDirectory);
+    }
+  }
+
+  public static void copy(File source, File destination) throws IOException,
+      InterruptedException {
+    Process process = new ProcessBuilder("cp", "-rp", source.getPath(),
+        destination.getPath()).start();
+
+    BufferedReader inputReader = new BufferedReader(new InputStreamReader(
+        process.getErrorStream()));
+
+    StringBuilder errorText = new StringBuilder();
+    String line = inputReader.readLine();
+
+    while (line != null) {
+      line = inputReader.readLine();
+      errorText.append(line);
+    }
+
+    if (process.waitFor() != 0) {
+      throw new IOException("Copy from " + source + " to " + destination
+          + " failed - " + errorText.toString());
+    }
+  }
+
+  @Override
+  public String toString() {
+    return String.format("{ workingDirectory:%s distributionDirectory:%s }",
+        workingDirectory, distributionDirectory);
   }
 
   public File getDistributionDirectory() {
